ENH cut MiniBatchKMeans memory usage in half for large n_clusters
Also, changed logging message to reflect the (public) class that is logging,
not the (private) function, and don't recompute distances in MBKMeans.
Previous code not only redid computations, but used 2 * n_clusters * batchsize
space to do it, orders of magnitude more than is required in the sparse case.

New Cython code lives in utils.sparsefuncs, because it may prove more
generally useful.

Fixes #1765.